created init package to avoid code repetition
